self assessment redirection issue
