+++ src/app/routes/self-assessment/self-assessment.component.html
@@ -1,1 +1,1 @@
-<lib-self-assessment (selfAsesment)="selfAsesment($event)"></lib-self-assessment>
+<lib-app-loader></lib-app-loader>
